backend/requirements.txt
fastapi
uvicorn
requests
instagrapi
pydantic
python-dotenv
sqlalchemy
aiosqlite
